requirement.txt
bitsandbytes==0.43.3
Cython==3.0.11
datasets==3.1.0
evaluate==0.4.3
flash-attn==2.6.3
flwr==1.7.0
learn2learn==0.1.7
matplotlib==3.9.2
peft==0.12.0
pip==24.2
platformdirs==4.2.2
promise==2.3
python-pptx==1.0.2
ray==2.35.0
rouge_score==0.1.2
sacrebleu==2.4.3
scikit-learn==1.5.1
sentencepiece==0.2.0
shortuuid==1.0.13
torchaudio==2.4.0
transformers==4.44.2
trl==0.10.1
unstructured==0.15.7
wandb==0.14.0
yaspin==3.0.2